refactor(core): Folderize workflow-sdk lint and validation modules
Group lint by sdk-builder vs code-node, and validation by node-parameter-schema vs node-port-resolvers, with thin barrels and more descriptive file names.
