+++ packages/@n8n/workflow-sdk/src/lint/code-node/extract-snippets.test.ts
@@ -1,4 +1,4 @@
-import { extractEmbeddedCodeSnippetsFromSource } from './extract-code-snippets';
+import { extractEmbeddedCodeSnippetsFromSource } from './extract-snippets';
 
 describe('extractEmbeddedCodeSnippetsFromSource', () => {
 	it('extracts jsCode from a Code node config', () => {
