+++ packages/@n8n/workflow-sdk/src/validation/node-parameter-schema/schema-helpers.test.ts
