+++ packages/@n8n/workflow-sdk/src/validation/node-parameter-schema/schema-helpers.ts
@@ -27,7 +27,7 @@ export {
 	literalUnion,
 	optionsWithExpression,
 	multiOptionsSchema,
-} from '../generate-types/zod-helpers';
+} from '../../generate-types/zod-helpers';
 
 // Re-export resolveSchema and types from resolve-schema
 export { resolveSchema, resolveOneOfSchemas } from './resolve-schema';
