+++ packages/@n8n/workflow-sdk/src/validation/node-parameter-schema/schema-validation-integration.test.ts
@@ -10,8 +10,8 @@
 
 import { validateNodeConfig, loadSchema } from './schema-validator';
 import { setupTestSchemas, teardownTestSchemas } from './test-schema-setup';
-import { parseWorkflowCode } from '../codegen/parse-workflow-code';
-import { validateWorkflow } from '../validation';
+import { parseWorkflowCode } from '../../codegen/parse-workflow-code';
+import { validateWorkflow } from '../../validation';
 
 function requireSchema(nodeType: string, version: number): void {
 	if (!loadSchema(nodeType, version)) {
