+++ packages/@n8n/workflow-sdk/src/validation/node-port-resolvers/resolve-main-input-count.ts
