+++ packages/@n8n/workflow-sdk/src/validation/output-index-validation.test.ts
