+++ packages/@n8n/workflow-sdk/src/validation/validate-workflow.test.ts
@@ -1,5 +1,5 @@
 import { validateWorkflow, ValidationError } from '.';
-import { setupTestSchemas, teardownTestSchemas } from './test-schema-setup';
+import { setupTestSchemas, teardownTestSchemas } from './node-parameter-schema/test-schema-setup';
 import type { NodeInstance, WorkflowJSON } from '../types/base';
 import { workflow } from '../workflow-builder';
 import { node, trigger, sticky } from '../workflow-builder/node-builders/node-builder';
